# fastingmate-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Dokumentasi API untuk autentikasi pengguna",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Menampilkan profil pengguna berdasarkan ID",
                "operationId": "getUserProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengguna yang ingin ditampilkan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "height": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 175
                                        },
                                        "weight": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 70
                                        },
                                        "activity_level": {
                                            "type": "string",
                                            "example": "moderate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Mengupdate profil pengguna berdasarkan ID",
                "operationId": "updateUserProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengguna yang ingin diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "height",
                                    "weight",
                                    "activity_level"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "height": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 175
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 70
                                    },
                                    "activity_level": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "moderate",
                                            "high"
                                        ],
                                        "example": "moderate"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "height": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 175
                                                },
                                                "weight": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 70
                                                },
                                                "activity_level": {
                                                    "type": "string",
                                                    "example": "moderate"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}/weight-category": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Menghitung kategori berat badan berdasarkan tinggi dan berat pengguna berdasarkan ID",
                "operationId": "getUserWeightCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengguna yang ingin dihitung kategori beratnya",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori berat badan berhasil dihitung",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "weight_category": {
                                            "type": "string",
                                            "example": "normal"
                                        },
                                        "bmi": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 22.86
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <your-token>)"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}